Changed startup scripts to reflect new configs

+++ runners/cp_cfg.py
@@ -11,21 +11,25 @@ def copy_config(id, ip):
 	host_dir = "/usr/local/nagios/etc/servers/"
 	host_dir += id
 	host_dir += ".cfg"
-	copyfile(old_file, host_dir)
-	print "config file copied!"
 	
-	for line in fileinput.input(host_dir, inplace=True):
-		print line.replace(minion, id),
-	print "hostname configured"
-	
-	for line in fileinput.input(host_dir, inplace=True):
-		print line.replace(ipaddress, ip),
-	print "ip configured"
-	
-	print "config complete, restarting nagios"
-	
-	os.system(restart)
-	
-	print "nagios restarted"
-	
+	if not (os.path.exists(host_dir))
+		copyfile(old_file, host_dir)
+		print "config file copied!"
+		
+		for line in fileinput.input(host_dir, inplace=True):
+			print line.replace(minion, id),
+		print "hostname configured"
+		
+		for line in fileinput.input(host_dir, inplace=True):
+			print line.replace(ipaddress, ip),
+		print "ip configured"
+		
+		print "config complete, restarting nagios"
+		
+		os.system(restart)
+		
+		print "nagios restarted"
+	else:
+		print "config exists, exiting"
+
 	return
